+++ Editor/Core/AssetDataStatus.cs
@@ -18,6 +18,7 @@ namespace Unity.AssetManager.Editor
         public static readonly AssetPreview.IStatus UploadSkip = new PreviewStatus(Constants.UploadSkipText, string.Empty, "grid-view--item-upload-skip");
         public static readonly AssetPreview.IStatus UploadOverride = new PreviewStatus(Constants.UploadOverrideText, string.Empty, "grid-view--item-upload-override");
         public static readonly AssetPreview.IStatus UploadDuplicate = new PreviewStatus(Constants.UploadDuplicateText, string.Empty, "grid-view--item-upload-duplicate");
+        public static readonly AssetPreview.IStatus UploadOutside = new PreviewStatus(Constants.UploadOutsideText, string.Empty, "grid-view--item-upload-outside");
 
         static class UssStyles
         {