+++ Editor/UI/Pages/AssetDetailsFooter.cs
@@ -66,10 +66,13 @@ namespace Unity.AssetManager.Editor
             m_ImportButton.SetEnabled(isEnabled);
 
             m_ShowInProjectBrowserButton.SetEnabled(enabled.HasFlag(UIEnabledStates.InProject));
-
+            m_ShowInProjectBrowserButton.tooltip = enabled.HasFlag(UIEnabledStates.InProject)
+                ? L10n.Tr(Constants.ShowInProjectButtonToolTip)
+                : L10n.Tr(Constants.ShowInProjectButtonDisabledToolTip);
+            
             m_RemoveImportButton.SetEnabled(enabled.HasFlag(UIEnabledStates.InProject) && !enabled.HasFlag(UIEnabledStates.IsImporting));
             m_RemoveImportButton.tooltip = enabled.HasFlag(UIEnabledStates.InProject)
-                ? string.Empty
+                ? L10n.Tr(Constants.RemoveFromProjectButtonToolTip)
                 : L10n.Tr(Constants.RemoveFromProjectButtonDisabledToolTip);
 
             m_OperationProgressBar.Refresh(operationInProgress);